+++ src/components/PeopleTable/PeopleTable.tsx
@@ -1,0 +1,64 @@
+import { Person } from '../../types';
+import { PersonalLink } from '../PersonalLink/PersonalLink';
+
+type Props = {
+  people: Person[];
+  slug?: string;
+};
+
+export const PeopleTable: React.FC<Props> = ({ people, slug }) => {
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {people.map(person => {
+          return (
+            <tr
+              data-cy="person"
+              className={person.slug === slug ? 'has-background-warning' : ''}
+              key={person.name}
+            >
+              <td>
+                <PersonalLink person={person} />
+              </td>
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                {person.mother ? (
+                  <PersonalLink person={person.mother} />
+                ) : person.motherName ? (
+                  person.motherName
+                ) : (
+                  '-'
+                )}
+              </td>
+              <td>
+                {person.father ? (
+                  <PersonalLink person={person.father} />
+                ) : person.fatherName ? (
+                  person.fatherName
+                ) : (
+                  '-'
+                )}
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
