+++ src/components/PersonalLink/PersonalLink.tsx
@@ -1,0 +1,17 @@
+import { Link } from 'react-router-dom';
+import { Person } from '../../types';
+
+type Props = {
+  person: Person;
+};
+
+export const PersonalLink: React.FC<Props> = ({ person }) => {
+  return (
+    <Link
+      className={person.sex === 'f' ? 'has-text-danger' : ''}
+      to={`/people/${person.slug}`}
+    >
+      {person.name}
+    </Link>
+  );
+};
